requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==2.0.15
pydantic==1.10.7
python-dotenv==1.0.0
redis==4.5.5
celery==5.3.1
python-multipart==0.0.6
passlib==1.7.4
python-jose==3.3.0
pyodbc
psycopg2-binary==2.9.7
openai==1.3.0
pytest==7.4.0
pytest-asyncio==0.21.1
httpx==0.24.1
